Renamed test refactoring class QuizFragmentsTestUsingDB
